+++ src/config.js
@@ -1,0 +1,8 @@
+export default {
+  red: 0xf25346,
+  white: 0xd8d0d1,
+  brown: 0x59332e,
+  pink: 0xF5986E,
+  brownDark: 0x23190f,
+  blue: 0x68c3c0,
+};
